feat: implement delete useCase
