+++ react-client/src/domain/repository/user/AccountEmergencyContactsRepository.ts
@@ -1,6 +1,7 @@
 import { AccountEmergencyContactModel } from "../../models/user/UserModels";
 
-export interface GetAccountEmergencyContactsRepository {
+export interface AccountEmergencyContactsRepository {
   query(): Promise<AccountEmergencyContactModel[] | []>;
   mutate(data: AccountEmergencyContactModel[] ): Promise<boolean>;
+  delete(data: AccountEmergencyContactModel ): Promise<boolean>;
 }
